fix: remove incorrect transpose in 2D contour plots
The pivot creates Z with shape (len(xb), len(xa)) which is correct for
contourf(xa, xb, Z). The .T transpose was flipping the axes, causing
the contour plot to be transposed relative to the star marker.

This resulted in the optimum star not overlapping with the minimum
loss region in the 2D loss surface plots.